fix(calendar): a ponta do intervalo ficava invisível no escuro, e a grade tinha 30 paradas de tabulação
Última rodada da revisão serial. Este é o componente que motivou a sonda
existir: dezesseis commits seguidos corrigiram um defeito por vez, cada um
relatado a olho, com a suíte verde o tempo todo. Desta vez a medição veio antes.

**Vue, tema escuro: ponta de intervalo em 1.18:1** — texto invisível. O gatilho
compunha `.nds-button-ghost` por fora, a lib marca o dia escolhido com
`aria-pressed`, e `.dark .nds-button-ghost[aria-pressed="true"]` tem
especificidade maior que a classe do calendário. **No claro os dois empatam e o
calendário vence por ordem de folha** — por isso só o escuro quebrava, e o axe
nunca olha o escuro.

**Vue: 30 paradas de tabulação na grade.** Uma correção anterior escrevia
`:tabindex="foraDoMes ? -1 : undefined"`, e `undefined` vindo do consumidor
APAGA a ligação interna da lib: todo dia do mês perdia o −1. A correção mirava
os dias vizinhos e desligou o roving tabindex inteiro. No Angular o mecanismo é
outro e o efeito parecido: `<button>` sem `tabindex` é tabulável, então dia
bloqueado e vizinho entravam na ordem.

`Home`, `End`, `PageUp` e `PageDown` não faziam nada em vue e svelte; no angular
viravam o mês e largavam o foco no `body`, porque a detecção de mudanças é
agendada e o foco acertava um botão da grade velha, destruída em seguida. A
aritmética virou módulo compartilhado, sem DOM.

React apagava o dia bloqueado duas vezes — opacidade na célula E no botão, 0.25
efetivo, 1.41:1 contra 2.10:1 nas outras quatro. E pintava a faixa de intervalo
com um par de tokens diferente das outras, no botão em vez da célula: no tema
default os dois coincidem e ninguém via.

O conteúdo compartilhado mentia em três idiomas: citava a lib de uma stack numa
página lida pelas cinco, prometia um papel ARIA que não existe em lugar nenhum,
cravava um atributo que duas stacks não emitem, e dizia que o foco ainda passa
pelo dia desabilitado — falso nas cinco.

O que foi medido e estava CERTO, que também é resultado: geometria idêntica ao
pixel nas cinco, papel de grade e 35 células, `aria-selected` sempre na célula,
rótulo do dia idêntico, abreviações e início de semana iguais, clique bloqueado
de verdade no dia desabilitado, e contraste do dia escolhido entre 6.35 e 17.18
nos três temas e dois modos.

Correção de método registrada no colhedor: a primeira medição de densidade
aplicou a classe num ancestral e deu os três degraus idênticos. Não era achado,
era régua errada — os degraus são declarados no `:root`, e custom property tem
o `var()` substituído onde é declarada. A célula, essa sim, tinha `2rem`
cravado e não acompanhava a densidade; passou a ler a escada de caixa, com o
degrau padrão byte-idêntico.

Uma regressão própria expôs um defeito latente no Vanilla: o passo de teclado
calculava a partir da data do botão e aplicava ao dia focado, que pode ser
outro — `Home` caía num sábado.

Fica declarado um buraco: o estado "hoje" não tem contraste medido, porque os
cenários são de mês fixo. Ele usa o mesmo par de tokens da forma que FOI
medida, mas a verificação direta continua sem número.

Suítes: react 77, vue 77, svelte 74, vanilla 80, angular 74. audit 0.

+++ nortear-design-system-react/src/components/ui/calendar.tsx
@@ -83,8 +83,13 @@ function Calendar({
       endMonth={fimPadrao}
       locale={locale}
       formatters={{
+        // Mês por EXTENSO no seletor, como no Vanilla, que é a referência: a
+        // forma curta em pt-BR sai com ponto ("jan."), e o ponto numa opção de
+        // uma palavra só é ruído — o mesmo motivo pelo qual ele já era removido
+        // do cabeçalho da semana. Medido, o seletor com o nome inteiro tem 61px
+        // contra 58px: cabe.
         formatMonthDropdown: (date) =>
-          date.toLocaleString(locale?.code, { month: "short" }),
+          date.toLocaleString(locale?.code, { month: "long" }),
         // Pelo Intl, e não pelos dados de locale da lib: ela devolve "sab" sem
         // acento, e as outras três stacks mostram "sáb". Mesma fonte nas
         // quatro, e sem o ponto que o pt-BR acrescenta em "dom.".
